+++ src/everest/bin/config_branch_script.py
@@ -50,7 +50,11 @@ def opt_controls_by_batch(optimization_dir: Path, batch: int) -> dict[str, Any] 
 
     assert storage is not None
     function_batch = next(
-        (b for b in storage.batches_with_function_results if b.batch_id == batch),
+        (
+            b
+            for b in storage.experiment.batches_with_function_results
+            if b.batch_id == batch
+        ),
         None,
     )
