+++ tests/everest/test_everserver.py
@@ -196,7 +196,7 @@ async def test_status_max_batch_num(copy_math_func_test_data_to_tmp):
     storage = EverestStorage.from_storage_path(config.storage_dir)
 
     # Check that there is only one batch.
-    assert {b.batch_id for b in storage.batches} == {0}
+    assert {b.batch_id for b in storage.experiment.batches} == {0}
 
 
 @pytest.mark.integration_test
